Be more strict about type conversion:   * numbers are no longer auto converted to booleans, use to_boolean or != 0   * booleans are no longer auto converted to numbers, use to_int   * strings will soon no longer be auto converted to numbers, use to_int
Added version information ("since 0.3.7") to documentation

git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@1006 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/script/scriptable.cpp
@@ -23,7 +23,7 @@ DECLARE_TYPEOF_COLLECTION(ScriptParseError);
 void store(const ScriptValueP& val, String& var)              { var = val->toString(); }
 void store(const ScriptValueP& val, int&    var)              { var = *val; }
 void store(const ScriptValueP& val, double& var)              { var = *val; }
-void store(const ScriptValueP& val, bool&   var)              { var = static_cast<int>(*val); }
+void store(const ScriptValueP& val, bool&   var)              { var = *val; }
 void store(const ScriptValueP& val, Color&  var)              { var = (AColor)*val; }
 void store(const ScriptValueP& val, AColor& var)              { var = *val; }
 void store(const ScriptValueP& val, Defaultable<String>& var) { var.assign(*val); }
